refactor: Moved Photos and Tasks to another file
